split agent runtime translation units
